add portal cooldown
portal cooldown:
New field (Lönn only) that makes the portal intangible for a certain amount of time. Due to the default of 0, it won't desync anything.

flag:
Added a flag "onPortalCooldown" + color (int) that is true when the portal is on cooldown. This is needed for the CC.

cleanup:
– only change portal colors when neccessary
– rename "teleportCooldown" to "teleportInsideCooldown" to distinguish to "cooldown"

Lönn cleanup:
Remove the separate placements for each color to reduce the number to just 5 placements (following issue #2 by Zucchini)